+++ src/api/cli/integration/integration_test.go
@@ -42,7 +42,7 @@ const (
 	// Number of random transactions of live transaction test.
 	randomLiveTransactionNum = 500
 
-	testFixturesDir = "test-fixtures"
+	testFixturesDir = "testdata"
 
 	stableWalletName        = "integration-test.wlt"
 	stableEncryptWalletName = "integration-test-encrypted.wlt"
@@ -104,8 +104,8 @@ func createEncryptedWallet(t *testing.T) (string, func()) {
 }
 
 // createTempWallet creates a temporary dir, and if encrypt is true, copy
-// the test-fixtures/$stableEncryptedWalletName file to the dir. If it's false, then
-// copy the test-fixtures/$stableWalletName file to the dir
+// the testdata/$stableEncryptedWalletName file to the dir. If it's false, then
+// copy the testdata/$stableWalletName file to the dir
 // returns the temporary wallet path, cleanup callback function, and error if any.
 func createTempWallet(t *testing.T, encrypt bool) (string, func()) {
 	dir, err := ioutil.TempDir("", "wallet-data-dir")
@@ -2036,23 +2036,23 @@ func TestStableCheckDB(t *testing.T) {
 	tt := []struct {
 		name   string
 		dbPath string
-		result []byte
-		errMsg []byte
+		result string
+		errMsg string
 	}{
 		{
 			name:   "no signature",
 			dbPath: "../../../visor/testdata/data.db.nosig",
-			errMsg: []byte("checkdb failed: find no signature of block: seq=1000\n"),
+			errMsg: "checkdb failed: Signature not found for block seq=1000 hash=71852c1a8ab5e470bd14e5fce8e1116697151181a188d4262b545542fb3d526c\n",
 		},
 		{
 			name:   "invalid database",
 			dbPath: "../../../visor/testdata/data.db.garbage",
-			errMsg: []byte("open db failed: invalid database\n"),
+			errMsg: "open db failed: invalid database\n",
 		},
 		{
 			name:   "valid database",
-			dbPath: "../../../gui/integration/test-fixtures/blockchain-180.db",
-			result: []byte("check db success\n"),
+			dbPath: "../../../gui/integration/testdata/blockchain-180.db",
+			result: "check db success\n",
 		},
 	}
 
@@ -2061,11 +2061,11 @@ func TestStableCheckDB(t *testing.T) {
 			output, err := exec.Command(binaryPath, "checkdb", tc.dbPath).CombinedOutput()
 			if err != nil {
 				fmt.Println(string(output))
-				require.Equal(t, tc.errMsg, output)
+				require.Equal(t, tc.errMsg, string(output))
 				return
 			}
 			require.NoError(t, err)
-			require.Equal(t, tc.result, output)
+			require.Equal(t, tc.result, string(output))
 		})
 	}
 }